+++ src/viewdom_wired/__init__.py
@@ -1,13 +1,11 @@
-from viewdom_wired.render import render, component, register_component
+from .protocols import Children, Component
+from .render import render, component, register_component
 
 __version__ = '0.1.0'
 
-
-def hello():
-    return 1
-
-
 __all__ = [
+    'Children',
+    'Component',
     'component',
     'register_component',
     'render',