+++ CHANGES.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+Unreleased
+==========
+
+- The ``Component`` protocol now indicates it returns a ``VDOM``
+
 0.1.0
 =====